requirements.txt
Flask
Flask-Mail
Flask-SQLAlchemy
Flask-WTF
Jinja2
MarkupSafe
SQLAlchemy
WTForms
Werkzeug
blinker
gunicorn
itsdangerous
virtualenv
wsgiref
uwsgi
Flask-kvsession
simplekv
pillow